creating service, repository and domain layers
