+++ bem-te-vi/service/auth_service.go
@@ -1,7 +1,8 @@
-package auth
+package service
 
 import (
 	"application"
+	"application/interfaces"
 	"context"
 	"fmt"
 	"strconv"
@@ -22,13 +23,13 @@ import (
 // }
 
 type AuthService struct {
-	userService application.UserService   
+	userService interfaces.UserService
 }
 
-func NewAuthService(userService application.UserService) *AuthService {
-    return &AuthService{
-        userService: userService,
-    }
+func NewAuthService(userService interfaces.UserService) *AuthService {
+	return &AuthService{
+		userService: userService,
+	}
 }
 
 func (s *AuthService) Login(ctx context.Context, query *application.LoginUserQuery) (*application.AuthInfo, error) {
@@ -44,7 +45,7 @@ func (s *AuthService) Login(ctx context.Context, query *application.LoginUserQue
 	}
 
 	return &application.AuthInfo{
-        AuthId: strconv.Itoa(user.ID),
-        Email:  user.Email,
-    }, nil
+		AuthId: strconv.Itoa(user.ID),
+		Email:  user.Email,
+	}, nil
 }
